Simplify FCM registration structure
Avoid registering on componentDidMount of ProjectList component. Instead, register only once - as indicated by a flag on the reducer 'fcm'. Also, change the responsibility of computing the token, from any view related component (or HOC) to fcmEpic. The epic decides when to and when not to update registration token. Also, add a refresh action of fcmTokens.
Doing all of this allows for complete removal of withFCMHandlers HOC in favor of a dispatch call in componentDidMount of ProjectListContainer.
This lays out the playing field for configuring foreground FCM notifications.